refactor: drain module-size ratchet to two monoliths (budget 1000)
Raise MODULE_SIZE_BUDGET from 900 to 1000 and drain six modules below it so
ALLOWLISTED_CEILINGS shrinks from eight entries to two (exceptions.py,
_artifacts.py — the genuine monoliths).

- _chat/api.py 955→899, _research.py 934→898, cli/source_cmd.py 964→957,
  _sources.py 1007→956, _artifact/downloads.py 1043→969: comment/docstring
  trims only (every issue/ADR ref and concurrency/RPC-shape rationale kept).
- _source/upload.py 1242→962: extract the pure URL-redaction, ADD_SOURCE_FILE
  source-id-extraction, and content-type helpers into the new sibling
  _source/_upload_decode.py, re-exported from upload.py so the
  notebooklm._source.upload.<helper> import/patch surface is preserved.
  Repoint the two urlsplit patch seams to the new module (same-PR).
- docs/architecture.md: list _source/_upload_decode.py in the repo tree + index.

No public-API change (audit_public_api_compat.py exits 0; allowlist untouched).
Ratchet, ruff, and mypy all green.
