+++ src/notebooklm/_artifact/downloads.py
@@ -64,38 +64,19 @@ async def _await_writer_exit(
 ) -> None:
     """Wait for a download writer thread to actually exit.
 
-    Plain ``await asyncio.to_thread(thread.join)`` is unsafe under
-    cancellation: if our awaiting task is cancelled, the await raises
-    ``CancelledError`` and we unwind even though the underlying
-    ``thread.join`` is still blocked on the thread. The thread keeps
-    running, which means the outer cleanup (``temp_file.unlink``)
-    races with the writer's still-open file handle.
+    A plain ``await asyncio.to_thread(thread.join)`` is unsafe under cancellation:
+    the await raises ``CancelledError`` and we unwind while the underlying join is
+    still blocked, so outer cleanup (``temp_file.unlink``) races the writer's
+    still-open file handle. ``asyncio.shield`` alone doesn't help (the await still
+    raises). The fix is a shield-loop that re-awaits the same shielded join task
+    until it completes; repeated cancellations only delay our re-raise, never the
+    writer's exit.
 
-    ``asyncio.shield`` alone doesn't fix this: it keeps the *inner*
-    join task alive across cancellation, but the *await* still raises
-    ``CancelledError`` and we unwind anyway. The fix is a shield-loop
-    that keeps re-awaiting the same shielded join task until it
-    actually completes. Repeated cancellations only delay our
-    re-raise, never the writer's exit.
-
-    Cancellation handling:
-
-    * Only ``asyncio.CancelledError`` is caught inside the loop — any
-      other exception from the shielded join (currently none in
-      practice, since ``Thread.join`` doesn't raise) propagates
-      immediately so we don't accidentally hide a real bug.
-    * The most recent ``CancelledError`` (if any) is preserved.
-    * If ``re_raise_cancel=True``, the helper re-raises that
-      ``CancelledError`` after the writer has fully exited. Callers
-      on the success path want this so an in-flight cancellation
-      isn't lost when the writer happens to finish first. Callers on
-      a cleanup-path (the producer's ``except`` block, which already
-      has an exception to re-raise) leave it at the default
-      ``False`` so we don't mask the original error with a second
-      cancellation.
-
-    Handles both the original join-vs-unlink race AND the follow-up case
-    where an initial fix could silently absorb task cancellation.
+    Only ``CancelledError`` is caught (any other join exception propagates). The
+    most recent ``CancelledError`` is preserved and, when ``re_raise_cancel`` is
+    set, re-raised after the writer exits — success-path callers want this so an
+    in-flight cancellation isn't lost; cleanup-path callers leave it ``False`` so
+    the original error isn't masked by a second cancellation.
     """
     join_task = asyncio.ensure_future(asyncio.to_thread(writer_thread.join))
     cancelled_error: asyncio.CancelledError | None = None
@@ -837,32 +818,16 @@ class ArtifactDownloadService:
                                 "Authentication may have expired. Run 'notebooklm login'.",
                             )
 
-                        # Producer/consumer split: a single dedicated
-                        # writer thread drains a bounded queue and writes
-                        # to ``temp_file``. Compared to the legacy
-                        # per-chunk ``asyncio.to_thread(f.write, chunk)``,
-                        # this avoids thousands of thread-pool allocations
-                        # for multi-GB downloads.
-                        #
-                        # The writer runs on a dedicated ``threading.Thread``
-                        # rather than ``asyncio.to_thread`` so it does NOT
-                        # tie up a slot in asyncio's default executor pool.
-                        # With many concurrent downloads, default-executor
-                        # saturation by long-lived writers (each blocking
-                        # on ``chunk_q.get()``) could deadlock producers
-                        # trying to ``put`` via ``to_thread``.
-                        #
-                        # Producer puts use ``put_nowait`` first and only
-                        # fall back to ``to_thread(put, ...)`` when the
-                        # queue is full, minimizing default-executor
-                        # pressure during normal flow.
-                        #
-                        # End-of-stream is signalled with a ``None``
-                        # sentinel. Writer-side failures are surfaced via
-                        # ``writer_error`` and an early ``writer_failed``
-                        # ``threading.Event`` so the producer can short-
-                        # circuit BEFORE the writer's drain completes,
-                        # avoiding wasted network reads.
+                        # Producer/consumer split: one dedicated ``threading.Thread``
+                        # (not ``asyncio.to_thread``, which would tie up a default-
+                        # executor slot and risk deadlocking producers under many
+                        # concurrent downloads) drains a bounded queue to
+                        # ``temp_file``, avoiding per-chunk thread-pool churn on
+                        # multi-GB files. Producer puts use ``put_nowait`` first,
+                        # falling back to ``to_thread(put)`` only when full. EOF is a
+                        # ``None`` sentinel; writer failures surface via
+                        # ``writer_error`` + an early ``writer_failed`` Event so the
+                        # producer can short-circuit before the drain completes.
                         chunk_q: queue.Queue[bytes | None] = queue.Queue(
                             maxsize=_DOWNLOAD_WRITER_QUEUE_SIZE
                         )
@@ -870,23 +835,12 @@ class ArtifactDownloadService:
                         writer_error: list[BaseException] = []
 
                         def _writer_loop() -> None:
-                            # If the writer raises (e.g. OSError on
-                            # ``fh.write``), the bounded queue may have a
-                            # producer parked in ``q.put`` waiting for a
-                            # consumer. Without draining, that producer
-                            # hangs forever because we are the only
-                            # consumer. The ``finally`` drains pending
-                            # items via ``get_nowait`` so blocked puts
-                            # complete and the producer can observe the
-                            # failure signal on its next iteration.
-                            #
-                            # ``writer_failed`` is set in the ``except``
-                            # BEFORE the drain so the producer's
-                            # short-circuit check fires as early as
-                            # possible — the drain itself can run for a
-                            # few milliseconds clearing the queue, during
-                            # which the producer would otherwise read and
-                            # discard network bytes pointlessly.
+                            # On writer failure the bounded queue may have a producer
+                            # parked in ``q.put``; the ``finally`` drains via
+                            # ``get_nowait`` so those puts complete and the producer
+                            # can observe the failure. ``writer_failed`` is set in
+                            # ``except`` BEFORE the drain so the producer short-
+                            # circuits as early as possible.
                             try:
                                 with open(temp_file, "wb") as fh:
                                     while True:
@@ -922,19 +876,13 @@ class ArtifactDownloadService:
                         try:
                             async for chunk in response.aiter_bytes(chunk_size=65536):
                                 if writer_failed.is_set():
-                                    # Writer raised mid-stream. Stop
-                                    # reading — further network bytes
-                                    # would just be discarded by the
-                                    # drain. The original error is
-                                    # re-raised via ``writer_error[0]``
-                                    # below.
+                                    # Writer raised mid-stream: stop reading (further
+                                    # bytes would just be drained); error re-raised
+                                    # via ``writer_error`` below.
                                     break
-                                # ``put_nowait`` avoids a ``to_thread``
-                                # round-trip when the queue has space
-                                # (the common case under balanced flow);
-                                # fall back to ``to_thread`` only when
-                                # the queue is full so the loop suspends
-                                # cleanly under back-pressure.
+                                # ``put_nowait`` avoids a ``to_thread`` round-trip
+                                # when the queue has space; fall back only when full
+                                # so the loop suspends cleanly under back-pressure.
                                 try:
                                     chunk_q.put_nowait(chunk)
                                 except queue.Full:
@@ -945,30 +893,20 @@ class ArtifactDownloadService:
                                     chunk_q.put_nowait(None)
                                 except queue.Full:
                                     await asyncio.to_thread(chunk_q.put, None)
-                            # Join surfaces any exception the writer
-                            # captured. ``_await_writer_exit`` shield-
-                            # loops until the writer actually exits so
-                            # the outer cleanup never races with the
-                            # still-open file handle. ``re_raise_cancel
-                            # =True`` ensures a cancellation that
-                            # arrived while we were waiting for the
-                            # writer isn't lost when the writer happens
-                            # to finish first.
+                            # ``_await_writer_exit`` shield-loops until the writer
+                            # exits (so cleanup never races its file handle) and
+                            # surfaces any captured exception; ``re_raise_cancel``
+                            # preserves a cancellation that arrived mid-wait.
                             await _await_writer_exit(writer_thread, re_raise_cancel=True)
                             if writer_error:
                                 raise next(iter(writer_error))  # one-slot exception box
                         except BaseException:
-                            # On producer-side failure (network error,
-                            # cancellation, HTML payload), make sure the
-                            # writer sees a sentinel and exits — even if
-                            # the queue is currently saturated. A bare
-                            # ``put_nowait(None)`` would raise
-                            # ``queue.Full`` and leave the writer parked
-                            # in ``q.get`` forever; instead drop one item
-                            # to make room, then put the sentinel. At
-                            # most two iterations are needed: the writer
-                            # is the only consumer, so once a slot opens
-                            # nothing else refills it.
+                            # On producer-side failure, ensure the writer sees a
+                            # sentinel and exits even if the queue is saturated: a
+                            # bare ``put_nowait(None)`` would raise ``queue.Full`` and
+                            # leave the writer parked forever, so drop one item to
+                            # make room then put the sentinel (≤2 iterations — the
+                            # writer is the only consumer).
                             while True:
                                 try:
                                     chunk_q.put_nowait(None)
@@ -978,23 +916,11 @@ class ArtifactDownloadService:
                                 try:
                                     chunk_q.get_nowait()
                                 except queue.Empty:
-                                    # Writer drained between our put and
-                                    # get — the next put attempt will
-                                    # succeed.
                                     pass
-                            # MUST wait for the writer to actually exit
-                            # before unwinding: the outer ``except``
-                            # unlinks ``temp_file``, which would race
-                            # with the writer's still-open file handle
-                            # otherwise. A plain
-                            # ``contextlib.suppress(BaseException) +
-                            # await to_thread(.join)`` does NOT suffice
-                            # — the await itself can be re-cancelled and
-                            # unwind before the writer finishes. The
-                            # shield-loop in ``_await_writer_exit``
-                            # keeps re-awaiting the same shielded join
-                            # task across repeated cancellations until
-                            # the writer thread is actually dead.
+                            # MUST wait for the writer to fully exit before
+                            # unwinding: the outer ``except`` unlinks ``temp_file``,
+                            # which would race the writer's open file handle. See
+                            # ``_await_writer_exit`` for why a plain join doesn't do.
                             await _await_writer_exit(writer_thread)
                             raise
