+++ src/notebooklm/_research.py
@@ -53,34 +53,13 @@ __all__ = [
 
 logger = logging.getLogger(__name__)
 
-# Sentinel marking "the canonical ``initial_interval`` keyword was not passed"
-# in ``wait_for_completion``. The default stays this ``object()`` sentinel (not a
-# literal ``5.0``) so the public-API compatibility audit's default-repr check
-# sees no changed-default break; when the caller leaves it unset we resolve the
-# cadence to ``_DEFAULT_RESEARCH_POLL_INTERVAL`` below.
+# Sentinel for "``initial_interval`` not passed" in ``wait_for_completion``. Kept
+# as ``object()`` (not literal ``5.0``) so the public-API compat default-repr
+# check sees no changed-default break; unset resolves to the default below.
 _INITIAL_INTERVAL_UNSET: Any = object()
 
-# Default poll cadence (seconds between status checks) used when
-# ``initial_interval`` is left unset. Preserved verbatim so default-shape callers
-# keep the same behavior.
+# Default poll cadence (seconds) when ``initial_interval`` is unset.
 _DEFAULT_RESEARCH_POLL_INTERVAL = 5.0
-
-
-# ---------------------------------------------------------------------------
-# IMPORT_RESEARCH timeout-verification helpers
-#
-# IMPORT_RESEARCH is classified NON_IDEMPOTENT_NO_RETRY in IDEMPOTENCY_REGISTRY
-# (see #808): the executor will surface the first 5xx/timeout to the caller
-# rather than retry blindly, because the wire protocol has no client-token
-# slot and a naive retry duplicates every source. ``ResearchAPI``'s
-# verification path sidesteps that constraint by snapshotting baseline
-# sources before the call and matching post-call ``sources.list`` URLs
-# against the request — disambiguating "server already committed but the
-# response was lost" from "request truly failed". These helpers mirror the
-# CLI-only logic that originally landed in PR #321 / #327; they live in the
-# library now so Python API consumers get the same deep-research fix the
-# CLI does (issue #315).
-# ---------------------------------------------------------------------------
 
 
 def _normalize_import_verification_url(url: str) -> str:
@@ -450,26 +429,23 @@ class ResearchAPI:
             await self._poll_task_models(notebook_id),
             notebook_id=notebook_id,
             task_id=task_id,
-            # The ambiguity raise only applies to the unfiltered (task_id is
-            # None) path; when a discriminator is pinned, _select_polled_tasks
-            # filters before the raise branch. Gating it here matches
-            # wait_for_completion and keeps the intent explicit.
+            # Ambiguity raise applies only to the unfiltered (task_id is None)
+            # path; a pinned discriminator filters before the raise. Matches
+            # wait_for_completion.
             raise_on_ambiguous=task_id is None,
         )
 
         if parsed_tasks:
-            # ``parsed_tasks`` is a typed ``list[ResearchTask]`` (not a decoded RPC
-            # payload); ``first_task, *_ = parsed_tasks`` avoids a ``name[int]`` read.
-            first_task, *_ = parsed_tasks  # typed list[ResearchTask] head; unpack avoids name[int]
+            # ``parsed_tasks`` is a typed ``list[ResearchTask]``; the unpack avoids
+            # a ``name[int]`` positional read on a decoded payload.
+            first_task, *_ = parsed_tasks
             return self._public_poll_result(first_task, parsed_tasks)
 
-        # A concrete pinned ``task_id`` that matched nothing is a poll-observed
-        # absence of that specific task — a typed ``NOT_FOUND`` sentinel
-        # carrying the requested id. A falsy ``task_id`` (``None`` for the
-        # unfiltered poll, or the degenerate empty string) is not a meaningful
-        # discriminator, so it stays ``NO_RESEARCH`` ("nothing in flight") and
-        # preserves the legacy empty-poll dict shape. See ADR-0019 Rule 4
-        # (#1346).
+        # A pinned ``task_id`` that matched nothing is a poll-observed absence —
+        # a typed ``NOT_FOUND`` sentinel carrying the id. A falsy ``task_id``
+        # (``None`` or empty string) is no discriminator, so it stays
+        # ``NO_RESEARCH`` and preserves the legacy empty-poll shape (ADR-0019
+        # Rule 4, #1346).
         if task_id:
             return ResearchTask.not_found(task_id)
 
@@ -526,11 +502,9 @@ class ResearchAPI:
                 positive.
             TypeError: If the resolved poll interval is not a number.
         """
-        # The sentinel default means "``initial_interval`` was not supplied" —
-        # fall back to the default cadence. An *explicit* non-numeric value
-        # (e.g. initial_interval=None or initial_interval="1") is a caller bug;
-        # fail fast with TypeError rather than silently coercing it back to the
-        # default.
+        # Unset sentinel → default cadence. An *explicit* non-numeric value
+        # (``None``, ``"1"``) is a caller bug: fail fast with TypeError rather
+        # than silently coercing it back to the default.
         if initial_interval is _INITIAL_INTERVAL_UNSET:
             poll_interval = _DEFAULT_RESEARCH_POLL_INTERVAL
         elif isinstance(initial_interval, bool) or not isinstance(initial_interval, (int, float)):
@@ -554,7 +528,6 @@ class ResearchAPI:
                 task_id=pinned_task_id,
                 raise_on_ambiguous=pinned_task_id is None,
             )
-            # ``parsed_tasks`` is a typed ``list[ResearchTask]``: first or ``None``.
             selected_task = next(iter(parsed_tasks), None)
             if pinned_task_id is None and selected_task is not None:
                 pinned_task_id = selected_task.task_id
@@ -683,8 +656,7 @@ class ResearchAPI:
         )
         imported = []
         # ``unwrap_import_rows`` centralises the ``[[src1, ...]]`` envelope probe
-        # (the former ``result[0]`` / ``first[0]`` reads) behind the research row
-        # adapter; an unrecognised shape falls through to ``[]``/unchanged.
+        # behind the research row adapter; an unrecognised shape → ``[]``.
         for src_data in unwrap_import_rows(result):
             row = ImportedSourceRow(src_data)
             if not row.is_well_formed:
@@ -834,22 +806,14 @@ class ResearchAPI:
                                 source_inputs, source_models, strict=True
                             )
                         ]
-                        # Filter for retry: drop already-present URLs.
-                        # Additionally, when *any* URL was verified
-                        # committed, drop no-URL entries (deep-research
-                        # reports): reports are appended FIRST in the
-                        # IMPORT_RESEARCH payload (see
-                        # ``_build_report_import_entry`` usage in
-                        # ``import_sources``), so a URL newly observed after
-                        # this attempt implies the report committed too.
-                        # Pre-existing URLs only de-dupe URL entries; they do
-                        # not prove this request committed no-URL reports.
-                        # Without this guard,
-                        # each retry duplicates the report server-side.
-                        # When no URL committed, keep no-URL entries —
-                        # the report's fate is unknown and the
-                        # report-only attempt cap further down bounds
-                        # the worst case.
+                        # Filter for retry: drop already-present URLs. Also, when
+                        # *any* URL committed, drop no-URL entries (deep-research
+                        # reports are appended FIRST in the IMPORT_RESEARCH payload,
+                        # so a newly-observed URL implies the report committed too —
+                        # without this guard each retry duplicates it server-side).
+                        # Pre-existing URLs only de-dupe URL entries. When no URL
+                        # committed, keep no-URL entries (report fate unknown; the
+                        # report-only attempt cap below bounds the worst case).
                         drop_no_url_entries = bool(committed_urls_norm)
                         filtered_source_pairs = [
                             (source_input, source)
